agent/requirements.txt
loguru>=0.7.2
pytest>=8.3
black>=24.8.0
ruff>=0.6.3
mypy>=1.11
pydantic>=2.7
httpx>=0.27
pyyaml>=6.0.1
typer>=0.12.5

# RAG
pdfminer.six==20240706
faiss-cpu==1.8.0.post1
sentence-transformers==2.7.0
